Rename context_manager to retarget
