+++ numba_dppy/__init__.py
@@ -519,7 +519,7 @@ import numba.testing
 
 from . import config
 
-from numba_dppy.context_manager import offload_to_sycl_device
+from numba_dppy.retarget import offload_to_sycl_device
 
 if config.dppy_present:
     from .device_init import *
